Add router utils folder
