+++ src/pages/saksbehandling/attestering/attesterKlage/AttesterKlage.tsx
@@ -2,7 +2,6 @@ import { Alert, Heading } from '@navikt/ds-react';
 import React from 'react';
 import { Link, useNavigate, useOutletContext } from 'react-router-dom';
 
-import { Person } from '~src/api/personApi';
 import { AttesteringsForm } from '~src/components/attestering/AttesteringsForm';
 import OppsummeringAvKlage from '~src/components/oppsummeringAvKlage/OppsummeringAvKlage';
 import * as klageActions from '~src/features/klage/klageActions';
@@ -10,23 +9,18 @@ import { useAsyncActionCreator } from '~src/lib/hooks';
 import { useI18n } from '~src/lib/i18n';
 import * as Routes from '~src/lib/routes';
 import { UnderkjennelseGrunn } from '~src/types/Behandling';
-import { Sak } from '~src/types/Sak';
 import {
     erKlageINoenFormForAvvist,
     erKlageOpprettholdt,
     erKlageTilAttestering,
     erKlageTilAttesteringAvvist,
 } from '~src/utils/klage/klageUtils';
+import { AttesteringContext } from '~src/utils/router/routerUtils';
 
 import * as sharedStyles from '../sharedStyles.module.less';
 
 import messages from './attesterKlage-nb';
 import * as styles from './attesterKlage.module.less';
-
-interface AttesteringContext {
-    sak: Sak;
-    søker: Person;
-}
 
 const AttesterKlage = () => {
     const { sak } = useOutletContext<AttesteringContext>();
