+++ src/utils/router/routerUtils.ts
@@ -1,0 +1,7 @@
+import { Person } from '~src/api/personApi';
+import { Sak } from '~src/types/Sak';
+
+export interface AttesteringContext {
+    sak: Sak;
+    søker: Person;
+}
